+++ crates/maple_cli/src/stdio_server/providers/filer.rs
@@ -103,6 +103,7 @@ impl EventHandler for FilerMessageHandler {
                     size: context.sensible_preview_size(),
                     context: &context,
                     inner: OnMove::Filer(path.clone()),
+                    expected_line: None,
                 };
                 if let Err(err) = on_move_handler.handle() {
                     log::error!("Failed to handle filer OnMove: {:?}", err);